requirements.txt
# 基础深度学习框架
pytorch>=2.0.0
pytorch-cuda=11.8  # 根据您的CUDA版本调整

# 模型和微调相关
transformers>=4.36.0
transformers_stream_generator>=0.0.4
torchaudio
torchvision
datasets>=2.10.0
peft>=0.8.0  # Parameter-Efficient Fine-Tuning
bitsandbytes>=0.41.0  # 用于量化
accelerate>=0.25.0  # 加速训练

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2
pyarrow>=12.0.0  # 高效数据读取

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 其他工具
tqdm>=4.65.0  # 进度条
loguru>=0.7.0  # 日志管理
sentencepiece>=0.1.99  # Tokenizer依赖

# Jupyter支持（可选，用于数据分析）
jupyterlab>=4.0.0
notebook>=7.0.0